+++ packages/subgraph/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UPCOMING]
 
+### Changed
+- Changes callId in `Executed` event from `uint256` to `bytes32`
+
 ## 0.6.2-alpha
 
 ### Added